backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "goodboi"
version = "0.1.0"
description = "AI-powered PRD generator that onboards to any project"
requires-python = ">=3.9"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-multipart>=0.0.12",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "rich>=13.0",
]

[project.scripts]
goodboi = "goodboi.cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["goodboi*"]

[tool.setuptools.package-data]
goodboi = ["static/**/*"]
